Revamp error reporting
This commit fixes #25 by converting the error reporting from
concatenated strings to lists of dictionaries, which contain the
error/warning code and the message.

The translated tests have been fixed to work with this, as well as
to fix other problems.

I have also rewritten most of the docstring type information to
reflect the actual Python type names.